@@ -1,4 +1,4 @@
-const sum = require('./sum');
+import sum from './sum';
 
 const total = sum(10, 5);
 console.log(total);
